add a whole bunch of tiles

+++ aerial_vg/tiles.py
@@ -2,6 +2,8 @@ import json
 import requests
 import os
 
+from progress.bar import ChargingBar 
+
 
 if __name__ == "__main__":
     with open("defs.json", 'r') as p:
@@ -13,24 +15,30 @@ if __name__ == "__main__":
 
         print(meta)
 
-        for lvl in levels: 
+        for idx, lvl in enumerate(levels): 
 
             rowMax = lvl["rowMax"]
             colMax = lvl["colMax"]
             zoom = lvl["level"]
 
-            print('zoom: {}, rows: {}, cols: {}'.format(zoom, rowMax, colMax))
+            bar = ChargingBar("Zoom Level: {}".format(zoom), max=rowMax * colMax)
 
             for x in range(colMax):
                 for y in range(rowMax):
 
-                    # get the tile
-                    url = 'http://base.maps.vic.gov.au/wmts/AERIAL_VG/EPSG:3111/{}/{}/{}.png'.format(zoom, x, y)
-                    r = requests.get(url, allow_redirects=True)
+                    writepath = 'tiles/{}/{}-{}.png'.format(zoom, x, y)
 
-                    writepath = 'tiles/{}-{}-{}.png'.format(zoom, x, y)
+                    if not os.path.exists(writepath):
 
-                    with open(writepath, 'wb') as q: 
-                        q.write(r.content)
+                        # get the tile
+                        url = 'http://base.maps.vic.gov.au/wmts/AERIAL_VG/EPSG:3111/{}/{}/{}.png'.format(zoom, x, y)
+                        r = requests.get(url, allow_redirects=True)
 
-            break
+                        
+                        # write content to file
+                        with open(writepath, 'wb') as q: 
+                            q.write(r.content)
+
+                    bar.next()
+
+            bar.finish()
